world: block support system
Now updated blocks fall when there isn't enough support from near blocks. This system isn't complete, as they should eventually land and change the ground or break to boulders, but that will be adressed later. Also this commit fixes (most of) exception which could crash the game when adding to/removing from ArrayLists

Signed-off-by: mkhropov <mkhropov@gmail.com>